requirements.txt
beautifulsoup4>=4.11.2
cachetools>=5.3.0
certifi>=2022.12.7
charset-normalizer>=3.0.1
click>=8.1.3
Flask>=2.2.3
google-api-core>=2.11.0
google-auth>=2.16.2
google-cloud-core>=2.3.2
google-cloud-datastore>=2.14.0
googleapis-common-protos>=1.58.0
grpcio>=1.51.3
grpcio-status>=1.51.3
gunicorn>=20.1.0
hjson>=3.1.0
icalendar>=5.0.4
idna>=3.4
itsdangerous>=2.1.2
Jinja2>=3.1.2
MarkupSafe>=2.1.2
proto-plus>=1.22.2
protobuf>=4.22.0
pyasn1>=0.4.8
pyasn1-modules>=0.2.8
python-dateutil>=2.8.2
pytz>=2022.7.1
requests>=2.28.2
rsa>=4.9
six>=1.16.0
soupsieve>=2.4
urllib3>=1.26.14
Werkzeug>=2.2.3
